+++ tests/Fixtures/PointInTimeTest.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types = 1);
 
-namespace Tests\Innmind\TimeContinuum\Fixtures\Earth;
+namespace Tests\Innmind\TimeContinuum\Fixtures;
 
-use Fixtures\Innmind\TimeContinuum\Earth\PointInTime;
+use Fixtures\Innmind\TimeContinuum\PointInTime;
 use Innmind\BlackBox\{
     Set,
     Random,